+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 
+- Support for PHP <7.3
+
 ## [2.0.0] - 2019-12-27
 
 ### Added