+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   now on `success` it returns data from `result` object preperty
   and on `error` it throws `Genius\Exception\ApiResponseErrorException`. 
   This means that return value and exceptions of **every** `Resource` has changed!
+- `Genius\ConnectGenius` is based on PSR-17 Discovery instead of deprecated `Http\Discovery\UriFactoryDiscovery`
 
 ### Removed